refactor(web): use only one Header for the whole app
