+++ packages/web/components/Header/Header.js
@@ -1,19 +1,23 @@
 import React from 'react';
-import { useTranslation } from 'react-i18next';
+import PropTypes from 'prop-types';
 import { useRouter } from 'next/router';
+import { useTranslation } from 'react-i18next';
 import styled, { css } from 'styled-components';
+import { Link as ScrollLink } from 'react-scroll';
 import { FaChevronDown } from 'react-icons/fa';
 import { Link } from '../Link';
-import links from './links';
+import { aboutLinks, defaultLinks } from './links';
 import { HamburguerMenu } from '../HamburguerMenu';
 import UserHeader from './UserHeader';
+import ShoppingCart from './ShoppingCart';
 import NewSolutionButton from './NewSolutionButton';
 import Dropdown from './Dropdown';
-import ShoppingCart from './ShoppingCart';
 
-const Header = () => {
+const Header = ({ isAbout }) => {
 	const { t } = useTranslation(['common']);
-	const { pathname } = useRouter();
+	const router = useRouter();
+	const links = isAbout ? aboutLinks : defaultLinks;
+
 	return (
 		<StyledHeader>
 			<Container>
@@ -23,30 +27,50 @@ const Header = () => {
 							<img src="/logo.svg" alt={t('common:logoDesc')} />
 						</Link>
 					</LogoContainer>
-					<MenuLinksWrapper>
-						<MenuLinksList>
-							{links.map(({ id, label, href, dropdown, external, sublinks = [] }) => (
-								<MenuLinksItem dropdown={dropdown} key={id}>
-									<Link
-										activeClass={pathname === href ? 'active' : ''}
-										href={href}
-										target={external ? '_blank' : '_self'}
-										rel={external ? 'noreferrer' : ''}
-									>
-										{label}
-										{dropdown && <FaChevronDown />}
-									</Link>
-									{dropdown && <Dropdown links={sublinks} />}
-								</MenuLinksItem>
-							))}
-						</MenuLinksList>
-					</MenuLinksWrapper>
 				</LeftContent>
 				<RightContent>
+					<MenuLinksWrapper>
+						<MenuLinksList>
+							{links.map(
+								({ id, label, to, href, dropdown, sublinks = [], scrollLink }) =>
+									scrollLink ? (
+										<MenuLinksItem dropdown={dropdown} key={id}>
+											<ScrollLink
+												activeClass="active"
+												to={to}
+												spy
+												smooth
+												duration={500}
+												offset={-65}
+											>
+												{label}
+												{dropdown && <FaChevronDown />}
+											</ScrollLink>
+											{dropdown && <Dropdown links={sublinks} />}
+										</MenuLinksItem>
+									) : (
+										<MenuLinksItem dropdown={dropdown} key={id}>
+											<Link
+												href={href}
+												className={
+													sublinks.find(
+														(link) => link.href === router.pathname,
+													) && 'active'
+												}
+											>
+												{label}
+												{dropdown && <FaChevronDown />}
+											</Link>
+											{dropdown && <Dropdown links={sublinks} />}
+										</MenuLinksItem>
+									),
+							)}
+						</MenuLinksList>
+					</MenuLinksWrapper>
 					<UserHeader />
 					<ShoppingCart />
 					<NewSolutionButton />
-					<HamburguerMenu links={links} scroll />
+					<HamburguerMenu links={links} secondary scroll />
 				</RightContent>
 			</Container>
 		</StyledHeader>
@@ -62,8 +86,7 @@ const StyledHeader = styled.header`
 		top: 0;
 		z-index: 1100;
 		background-color: ${colors.white};
-		box-shadow: 0 0.1rem 0.3rem ${colors.darkWhite};
-		border-bottom: 0.1rem solid ${colors.border};
+		box-shadow: 0 0 0.4rem ${colors.darkGray3};
 
 		@media (max-width: ${screens.medium}px) {
 			height: 6rem;
@@ -84,26 +107,19 @@ const LeftContent = styled.div`
 `;
 
 const LogoContainer = styled.div`
-	${({ theme: { colors, screens } }) => css`
-		border-right: 0.1rem solid ${colors.border};
-		width: 100%;
+	width: 100%;
+	height: 100%;
+
+	img {
 		height: 100%;
-
-		img {
-			height: 100%;
-			width: 100%;
-		}
-
-		@media (max-width: ${screens.medium}px) {
-			border-right: none;
-		}
-	`}
+		width: 100%;
+	}
 `;
 
 const MenuLinksWrapper = styled.nav`
-	${({ theme: { screens } }) => css`
-		height: 100%;
+	height: 100%;
 
+	${({ theme: { screens } }) => css`
 		@media (max-width: ${screens.large}px) {
 			display: none;
 		}
@@ -118,16 +134,21 @@ const MenuLinksList = styled.ul`
 `;
 
 const MenuLinksItem = styled.li`
-	${({ theme: { colors, screens } }) => css`
+	${({ theme: { colors } }) => css`
 		font-size: 1.5rem;
+		height: 100%;
 		display: flex;
 		align-items: center;
+		padding: 0 1.2rem;
 		position: relative;
 
 		a {
-			padding: 0 3rem;
+			padding: 0;
+			font-weight: bold;
 			text-transform: uppercase;
 			color: ${colors.black};
+			border-bottom: 0.4rem solid transparent;
+			height: 100%;
 			display: inline-flex;
 			justify-content: center;
 			align-items: center;
@@ -141,8 +162,8 @@ const MenuLinksItem = styled.li`
 				border-color: ${colors.secondary};
 			}
 
-			@media (max-width: ${screens.huge}px) {
-				padding: 0 1rem;
+			svg {
+				margin-left: 0.8rem;
 			}
 		}
 
@@ -161,5 +182,13 @@ const RightContent = styled.div`
 	display: flex;
 	align-items: center;
 `;
+
+Header.propTypes = {
+	isAbout: PropTypes.bool,
+};
+
+Header.defaultProps = {
+	isAbout: false,
+};
 
 export default Header;
