+++ packages/web/components/Header/links.js
@@ -1,84 +1,132 @@
-const links = [
+export const defaultLinks = [
 	{
 		id: 1,
-		label: 'Início',
-		href: '/',
+		label: 'Plataforma',
+		href: '/about',
 		dropdown: false,
-		external: false,
+		scrollLink: false,
 	},
 	{
 		id: 2,
-		label: 'Busca',
-		href: '/search',
+		label: 'Funcionalidades',
+		href: '/about#features',
 		dropdown: false,
-		external: false,
+		scrollLink: false,
 	},
 	{
 		id: 3,
 		label: 'Recursos',
-		href: '/about/#resources',
+		href: '/about#resources',
 		dropdown: true,
-		external: false,
+		scrollLink: false,
 		sublinks: [
 			{
 				id: 1,
 				label: 'Cursos',
 				href: '/courses',
-				external: false,
 			},
 			{
 				id: 2,
 				label: 'Fórum',
 				href: '/forum',
-				external: false,
 			},
 			{
 				id: 3,
 				label: 'Ajuda',
 				href: '/help',
-				external: false,
 			},
 			{
 				id: 4,
 				label: 'Banco de Ideias',
 				href: '/ideas-bank',
-				external: false,
 			},
 			{
 				id: 5,
 				label: 'Banco de Editais',
 				href: '/announcements-bank',
-				external: false,
 			},
 			{
 				id: 6,
 				label: 'Banco de Pesquisadores',
 				href: '/researchers-bank',
-				external: false,
 			},
 		],
 	},
 	{
 		id: 4,
-		label: 'Sobre',
-		href: '/about',
-		dropdown: false,
-		external: false,
-	},
-	{
-		id: 5,
 		label: 'Contato',
-		href: '/contact-us',
+		href: '/about#contact',
 		dropdown: false,
-		external: false,
-	},
-	{
-		id: 6,
-		label: 'Blog',
-		href: 'https://blog.plataformasabia.com/',
-		dropdown: false,
-		external: true,
+		scrollLink: false,
 	},
 ];
 
-export default links;
+export const aboutLinks = [
+	{
+		id: 0,
+		label: 'Início',
+		href: '/',
+		dropdown: false,
+		scrollLink: false,
+	},
+	{
+		id: 1,
+		label: 'Plataforma',
+		to: 'intro',
+		dropdown: false,
+		scrollLink: true,
+	},
+	{
+		id: 2,
+		label: 'Funcionalidades',
+		to: 'features',
+		dropdown: false,
+		scrollLink: true,
+	},
+	{
+		id: 3,
+		label: 'Recursos',
+		to: 'resources',
+		dropdown: true,
+		scrollLink: true,
+		sublinks: [
+			{
+				id: 1,
+				label: 'Cursos',
+				href: '/courses',
+			},
+			{
+				id: 2,
+				label: 'Fórum',
+				href: '/forum',
+			},
+			{
+				id: 3,
+				label: 'Ajuda',
+				href: '/help',
+			},
+			{
+				id: 4,
+				label: 'Banco de Ideias',
+				href: '/ideas-bank',
+			},
+			{
+				id: 5,
+				label: 'Banco de Editais',
+				href: '/announcements-bank',
+			},
+			{
+				id: 6,
+				label: 'Banco de Pesquisadores',
+				href: '/researchers-bank',
+			},
+		],
+	},
+	{
+		id: 4,
+		label: 'Contato',
+		to: 'contact',
+		dropdown: false,
+		scrollLink: true,
+	},
+];
