+++ packages/web/components/_Layouts/LandingPage.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { useRouter } from 'next/router';
 import Head from '../head';
-import { Header } from '../LandingPage/Header';
+import { Header } from '../Header';
 import { Footer } from '../Footer';
 
 const LayoutLandingPage = ({ children }) => {
